fix: correct link to user profile
If we located in `Stream Manager` and opening user card, with origin we going to `dashboard.twitch.tv/{username}`, which is wrong link.

So, i hardcoded twitch url to correctly handle this.

I'm not sure does this component used by youtube, if yes tell me i'll add checks for site address in this function.