fix(twitch): replies in thread (#1007)
Opening / replying to threads would only allow you to view up to a maximum of 2 messages (the main message and one reply).
This is fixed by using the tray's thread props instead of just it's parent props. This PR also fixes the issue where message buttons (copy & reply) were not being displayed when hovered over inside of a thread tray.